fix of self.y_l (rounding) PointPillarsPollo point_cloud_range fix

+++ model/pointpillars_pollo.py
@@ -89,7 +89,7 @@ class PillarEncoder(nn.Module):
         features *= mask[:, :, None]
 
         # 5. embedding
-        features = features.permute(0, 2, 1).contiguous()  # (p1 + p2 + ... + pb, 8, num_points)
+        features = features.permute(0, 2, 1).contiguous().float()  # (p1 + p2 + ... + pb, 8, num_points)
         features = F.relu(self.bn(self.conv(features)))  # (p1 + p2 + ... + pb, out_channels, num_points)
         pooling_features = torch.max(features, dim=-1)[0]  # (p1 + p2 + ... + pb, out_channels)
